+++ NEWS.md
@@ -1,6 +1,7 @@
 # deformula 0.1.2
 
-* The package is refactored with Roxygen and devtools.
-* License has been changed.
+- The package has been rewritten with Rcpp, roxygen2, testthat and devtools.
+- License has been changed.
+